refactor(core): centralize observability composition
