+++ packages/core/src/observability/shared.ts
@@ -1,0 +1,1 @@
+export const runID = crypto.randomUUID()
